requirements.txt
spacy>=3.0.6,<4.0.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0.tar.gz#egg=en_core_web_sm

Pillow
PyMuPDF
pytesseract
PyPDF2

streamlit
langchain
openai